# Fix issue with coding style fixer failing when list of files is too long