At least fails for me on a stock Linux system. 

Excerpt from `xargs --show-limits`
```
POSIX upper limit on argument length (this system): 16772719
POSIX smallest allowable upper limit on argument length (all systems): 4096
Maximum length of command we could actually use: 16770270
```