@mikee47 2022-02-02
Is there a performance hit with this?