@mikee47 2022-02-02
> > Is there a performance hit with this?
> 
> Yes, for every file in the list a new shell will be opened. But the delay is less than 30 seconds (at least on my Ubuntu box). Other possibility would be to split the big list of files and to pass 20 files at a time to the `clang-format` app. I just don't have enough makefile `fu` skills to do it.

Try this... should be a teeny bit quicker :-)