@slaff 2022-02-02
> > > Is there a performance hit with this?
> > 
> > 
> > Yes, for every file in the list a new shell will be opened. But the delay is less than 30 seconds (at least on my Ubuntu box). Other possibility would be to split the big list of files and to pass 20 files at a time to the `clang-format` app. I just don't have enough makefile `fu` skills to do it.
> 
> Try this... should be a teeny bit quicker :-)

Wonderful. It is indeed much faster. Here are the results on my machine: 

Before:

```sh
slaff@computer:/media/sla....box/dev/Sming/Sming/$ time make cs
make: Nothing to be done for 'cs'.

real	0m24,658s
user	0m12,166s
sys	0m8,524s
slaff@computer:/media/sla....box/dev/Sming/Sming/$ time make cs
make: Nothing to be done for 'cs'.

real	0m20,467s
user	0m11,879s
sys	0m8,286s
slaff@computer:/media/sla....box/dev/Sming/Sming/$ time make cs
make: Nothing to be done for 'cs'.

real	0m20,843s
user	0m12,052s
sys	0m8,366s
slaff@computer:/media/sla....box/dev/Sming/Sming/$
```

After:

```
slaff@computer:/media/sla....box/dev/Sming/Sming/$ time make cs

real	0m5,611s
user	0m4,748s
sys	0m0,757s
slaff@computer:/media/sla....box/dev/Sming/Sming/$ time make cs

real	0m5,609s
user	0m4,775s
sys	0m0,746s
slaff@computer:/media/sla....box/dev/Sming/Sming/$ time make cs

real	0m5,597s
user	0m4,669s
sys	0m0,844s
```

@mikee47 I will merge this PR if you don't have other optimizations in mind.